+++ .env
@@ -15,7 +15,7 @@ HASH_OPENSEARCH_USERNAME=admin
 HASH_AGENT_RUNNER_ENABLED=true
 HASH_AGENT_RUNNER_HOST=127.0.0.1
 HASH_AGENT_RUNNER_PORT=7123
-HASH_AGENT_RUNNER_LOG_LEVEL=INFO
+HASH_AGENT_RUNNER_LOG_LEVEL=DEBUG
 
 POSTGRES_PORT=5432
