+++ apps/hash-agents/app/logger/__init__.py
@@ -1,0 +1,4 @@
+from .http import logging_middleware as http_logging_middleware
+from .setup import Environment, setup_logging
+
+__all__ = ["Environment", "setup_logging", "http_logging_middleware"]
