+++ apps/hash-agents/app/routes/health.py
@@ -3,6 +3,6 @@ from fastapi import APIRouter, Response
 router = APIRouter()
 
 
-@router.get("/")
+@router.get("")
 async def health() -> Response:
     return Response(status_code=200)
